// packages/story-player/src/rn/types.ts
/*
 * Copyright 2026 Google LLC
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     https://www.apache.org/licenses/LICENSE-2.0
 */

export interface RNPageSize {
  width: number;
  height: number;
}

export interface RNBox {
  x: number;
  y: number;
  width: number;
  height: number;
}